User entity in progress
