+++ src/model/authetication.ts
@@ -1,0 +1,3 @@
+type authenticationData = {
+    id: string
+}
