+++ src/model/postTypeEnum.ts
@@ -1,4 +1,4 @@
 enum postType {
     NORMAL = 'Normal',
-    EVENTO = 'Evento'
+    EVENT = 'Event'
 }
