+++ src/services/idGenerator.ts
@@ -1,0 +1,5 @@
+import { v4 } from 'uuid'
+
+export const generateId = () => {
+    return v4()
+}
